[REBASE] Añadir endpoints y operaciones CRUD para las compras

+++ db/definiciones/calificaciones_tienda.sql
@@ -1,7 +1,7 @@
 CREATE TABLE calificaciones_tienda (
 	id_store BIGINT(20) NOT NULL,
 	id_user BIGINT(20) NOT NULL,
-	score TINYINT(1) NOT NULL,
+	score DECIMAL(2,1) NOT NULL,
 	`description` VARCHAR(255),
 	PRIMARY KEY (id_user, id_store)
 );
